Updated Skewness-Calculator script
This version of the script calculates the Pearson coefficient of skewness using only the values for contigs where reads mapped, not all contigs as before. This results in a more accurate skew value, whereas there was a bias associated with number of contigs in the previous version (many contigs = dilution of the real skewness signal).